+++ transform/txformConstants.go
@@ -1,0 +1,9 @@
+package transform
+
+const (
+	Vendor              string = "vendor"
+	MonitoredObjectType string = "objectType"
+	Metric              string = "metric"
+	Direction           string = "direction"
+	MonitoredObjectIds  string = "monitoredObjectIds"
+)